+++ CppCurl/CppCurl.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.30225.117
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CppCurl", "CppCurl.vcxproj", "{CAC19F5E-2D10-473A-8C9D-C41A3FB3136A}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CppCurlDLL", "..\CppCurlDLL\CppCurlDLL.vcxproj", "{AE768816-F923-46A3-B548-2FABBC793472}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{CAC19F5E-2D10-473A-8C9D-C41A3FB3136A}.Release|x64.Build.0 = Release|x64
 		{CAC19F5E-2D10-473A-8C9D-C41A3FB3136A}.Release|x86.ActiveCfg = Release|Win32
 		{CAC19F5E-2D10-473A-8C9D-C41A3FB3136A}.Release|x86.Build.0 = Release|Win32
+		{AE768816-F923-46A3-B548-2FABBC793472}.Debug|x64.ActiveCfg = Debug|x64
+		{AE768816-F923-46A3-B548-2FABBC793472}.Debug|x64.Build.0 = Debug|x64
+		{AE768816-F923-46A3-B548-2FABBC793472}.Debug|x86.ActiveCfg = Debug|Win32
+		{AE768816-F923-46A3-B548-2FABBC793472}.Debug|x86.Build.0 = Debug|Win32
+		{AE768816-F923-46A3-B548-2FABBC793472}.Release|x64.ActiveCfg = Release|x64
+		{AE768816-F923-46A3-B548-2FABBC793472}.Release|x64.Build.0 = Release|x64
+		{AE768816-F923-46A3-B548-2FABBC793472}.Release|x86.ActiveCfg = Release|Win32
+		{AE768816-F923-46A3-B548-2FABBC793472}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
