+++ CppCurlDLL/CppCurl.h
