+++ CppCurlDLL/Encoder.h
